requirements.txt
numpy>=1.21.0
scipy>=1.7.0
cvxpy>=1.2.0
matplotlib>=3.4.0
pyyaml>=5.4.0
tqdm>=4.62.0
scikit-learn>=1.0.0

